# services/tracking-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the WR any% automatic route tracking application.\nWebSocket tracker: ws://host/v1/ws/tracker — see docs/contracts/websocket-tracker-protocol.md",
        "title": "WR any% API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/tracking/points": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracking"
                ],
                "summary": "List raw GPS points for the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by device UUID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start of time range (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of time range (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results per page (default 1000, max 5000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor from previous response",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TrackingPointsEnv"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
        "/v1/tracking/points/{event_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tracking"
                ],
                "summary": "Delete a single GPS point by event_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event UUID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
        "/v1/tracking/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracking"
                ],
                "summary": "Get aggregated stats for a time range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by device UUID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start of time range (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of time range (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TrackingSummaryEnv"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
        "/v1/tracking/track": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracking"
                ],
                "summary": "Get simplified track: stationary clusters collapsed to a centroid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by device UUID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start of time range (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End of time range (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TrackEnv"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.ApiError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
...
        "http.TrackEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.TrackResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "http.TrackResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.TrackSegmentItem"
                    }
                }
            }
        },
        "http.TrackSegmentItem": {
            "type": "object",
            "properties": {
                "accuracy_m": {
                    "type": "number"
                },
                "event_id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string",
                    "example": "stay"
                },
                "lat": {
                    "type": "number",
                    "example": 55.751244
                },
                "lon": {
                    "type": "number",
                    "example": 37.618423
                },
                "merged_count": {
                    "type": "integer",
                    "example": 180
                },
                "period_end": {
                    "type": "string",
                    "example": "2026-06-10T14:30:00Z"
                },
                "recorded_at": {
                    "type": "string",
                    "example": "2026-06-10T12:00:01Z"
                },
                "segment_id": {
                    "type": "integer",
                    "example": 2
                },
                "speed_mps": {
                    "type": "number"
                },
                "stay_duration_sec": {
                    "type": "integer",
                    "example": 9000
                }
            }
        },
        "http.TrackingPoint": {
            "type": "object",
            "properties": {
                "accuracy_m": {
                    "type": "number",
                    "example": 8.5
                },
                "activity_type": {
                    "type": "string",
                    "example": "walking"
                },
                "bearing_deg": {
                    "type": "number"
                },
                "device_id": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "lat": {
                    "type": "number",
                    "example": 55.751244
                },
                "lon": {
                    "type": "number",
                    "example": 37.618423
                },
                "recorded_at": {
                    "type": "string",
                    "example": "2026-06-10T12:00:01Z"
                },
                "speed_mps": {
                    "type": "number"
                }
            }
        },
        "http.TrackingPointsEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.TrackingPointsResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "http.TrackingPointsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.TrackingPoint"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "http.TrackingSummary": {
            "type": "object",
            "properties": {
                "avg_speed_mps": {
                    "type": "number"
                },
                "duration_sec": {
                    "type": "integer",
                    "example": 1800
                },
                "first_recorded_at": {
                    "type": "string",
                    "example": "2026-06-10T12:00:01Z"
                },
                "last_recorded_at": {
                    "type": "string",
                    "example": "2026-06-10T12:30:01Z"
                },
                "max_speed_mps": {
                    "type": "number"
                },
                "points_count": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "http.TrackingSummaryEnv": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.TrackingSummary"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT access token. For WebSocket upgrade ?access_token=\u003cjwt\u003e is also accepted.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}